+++ test_package/test_atmos.cpp
@@ -32,11 +32,11 @@ void testTrueAirspeed1(double ias, double altFt, double satC){
     cout << "\tMach: \t\t" << M << endl;
     cout << "\tEAS: \t\t" << MathUtil::convertMpersToKts(eas) << "kts" << endl;
 
-    double geoTas = GeoUtil::convertIasToTas(ias, GeoUtil::STD_PRES_HPA, altFt, MathUtil::convertKelvinToCelsius(T));
+    //double geoTas = GeoUtil::convertIasToTas(ias, GeoUtil::STD_PRES_HPA, altFt, MathUtil::convertKelvinToCelsius(T));
 
     cout << "\tOnline TAS: \t" << onlineTas << "kts" << endl;
     cout << "\tAtmos. TAS: \t" << MathUtil::convertMpersToKts(atmosTas) << "kts" << endl;
-    cout << "\tGeo. TAS: \t" << geoTas << "kts" << endl;
+    //cout << "\tGeo. TAS: \t" << geoTas << "kts" << endl;
 }
 
 void testConvertTasToIas(){
@@ -48,7 +48,14 @@ void testConvertTasToIas(){
     cout << "\tActually: " << ias << endl;
 }
 
+void testCalculatePressureAtAlt(){
+    cout << "Test calculatePressureAtAlt 1:" << endl;
+    double P = AtmosUtil::calculatePressureAtAlt(10, 0, 1013.25, 273.15 + 25);
+    cout << "\tP is " << P << endl;
+}
+
 int main(){
     testTrueAirspeed1(260, 35000, -39);
     testConvertTasToIas();
+    testCalculatePressureAtAlt();
 }